+++ src/slic3r/GUI/AboutDialog.cpp
@@ -1,6 +1,7 @@
 #include "AboutDialog.hpp"
+#include "I18N.hpp"
 
-#include "../../libslic3r/Utils.hpp"
+#include "libslic3r/Utils.hpp"
 
 namespace Slic3r { 
 namespace GUI {
