+++ src/slic3r/GUI/ButtonsDescription.cpp
@@ -6,6 +6,7 @@
 
 #include "GUI.hpp"
 #include "GUI_App.hpp"
+#include "I18N.hpp"
 
 namespace Slic3r {
 namespace GUI {
