+++ src/slic3r/GUI/I18N.cpp
@@ -1,0 +1,11 @@
+#include "I18N.hpp"
+
+namespace Slic3r { namespace GUI { 
+
+wxString L_str(const std::string &str)
+{
+	//! Explicitly specify that the source string is already in UTF-8 encoding
+	return wxGetTranslation(wxString(str.c_str(), wxConvUTF8));
+}
+
+} }
