+++ src/slic3r/GUI/Preferences.cpp
@@ -1,6 +1,7 @@
 #include "Preferences.hpp"
 #include "AppConfig.hpp"
 #include "OptionsGroup.hpp"
+#include "I18N.hpp"
 
 namespace Slic3r {
 namespace GUI {
