+++ src/slic3r/GUI/PresetBundle.cpp
@@ -1,9 +1,9 @@
-//#undef NDEBUG
 #include <cassert>
 
 #include "PresetBundle.hpp"
 #include "BitmapCache.hpp"
 #include "Plater.hpp"
+#include "I18N.hpp"
 
 #include <algorithm>
 #include <fstream>
@@ -25,9 +25,9 @@
 #include <wx/bmpcbox.h>
 #include <wx/wupdlock.h>
 
-#include "../../libslic3r/libslic3r.h"
-#include "../../libslic3r/PlaceholderParser.hpp"
-#include "../../libslic3r/Utils.hpp"
+#include "libslic3r/libslic3r.h"
+#include "libslic3r/PlaceholderParser.hpp"
+#include "libslic3r/Utils.hpp"
 
 // Store the print/filament/printer presets into a "presets" subdirectory of the Slic3rPE config dir.
 // This breaks compatibility with the upstream Slic3r if the --datadir is used to switch between the two versions.
