+++ src/slic3r/Utils/PrintHostSendDialog.cpp
@@ -10,7 +10,7 @@
 
 #include "slic3r/GUI/GUI.hpp"
 #include "slic3r/GUI/MsgDialog.hpp"
-
+#include "slic3r/GUI/I18N.hpp"
 
 namespace fs = boost::filesystem;
